--- utils/generateEnv.js
@@ -1,5 +1,0 @@
-const generateEnv = () => {
-  return `SKIP_PREFLIGHT_CHECK=true`
-}
-
-module.exports = generateEnv;
