--- utils/generateJSON.js
@@ -1,5 +1,0 @@
-function generateJSON(data) {
-  return data
-}
-
-module.exports = generateJSON;
