--- utils/generateModelsIndex.js
@@ -1,8 +1,0 @@
-const generateModelsIndex = () => {
-  return `const User = require('./User');
-
-module.exports = { User }  
-  `
-}
-
-module.exports = generateModelsIndex;
