--- utils/generateSchemasIndex.js
@@ -1,9 +1,0 @@
-const generateSchemasIndex = () => {
-  return `const typeDefs = require('./typeDefs');
-const resolvers = require('./resolvers');
-
-module.exports = { typeDefs, resolvers };
-  `
-}
-
-module.exports = generateSchemasIndex
