+++ utils/generators.js
@@ -1,0 +1,430 @@
+export const generateAuth = () => {
+  return `const jwt = require("jsonwebtoken");
+
+const secret = "8675309didyoueverhearthetragedyofdarthplagueisthewisePLEASEputYOURownSECREThere";
+const expiration = "24h";
+
+module.exports = {
+  signToken: function ({ username, email, _id }) {
+    const payload = { username, email, _id };
+
+    return jwt.sign({ data: payload }, secret, { expiresIn: expiration });
+  },
+  authMiddleware: function({ req }) {
+    // permits token to be sent via any of the below
+    let token = req.body.token || req.query.token || req.headers.authorization;
+  
+    // separate "Bearer" from the value of the token itself
+    if (req.headers.authorization) {
+      token = token
+        .split(' ')
+        .pop()
+        .trim();
+    }
+  
+    // if no token has been provided, return request object as is
+    if (!token) {
+      return req;
+    }
+  
+    try {
+      // decode and attach user data to request object
+      const { data } = jwt.verify(token, secret, { maxAge: expiration });
+      req.user = data;
+    } catch {
+      console.log('This token is invalid. You shall not pass.');
+    }
+  
+    // return updated request object
+    return req;
+  }
+};
+  `
+}
+
+export const generateConnection = (response) => {
+  // js to format db name based on project name
+  let dbName = response.title.toLowerCase()
+  dbName = dbName.replace(/\s+/g, '-');
+
+  return `const mongoose = require('mongoose');
+
+mongoose.connect(process.env.MONGODB_URI || 'mongodb://localhost/${dbName}', {
+  useNewUrlParser: true,
+  useUnifiedTopology: true
+});
+
+module.exports = mongoose.connection; 
+  `
+}
+
+export const generateServer = () => {
+  return `const express = require('express');
+const path = require('path');
+
+// import ApolloServer
+const { ApolloServer } = require('apollo-server-express');
+const { authMiddleware } = require('./utils/auth');
+
+// import typeDefs and resolvers
+const { typeDefs, resolvers } = require('./schemas');
+
+const db = require('./config/connection');
+
+const PORT = process.env.PORT || 3001;
+const app = express();
+
+// declare startServer function
+const startServer = async () => {
+  // establish Apollo server with data from ./schemas
+  const server = new ApolloServer({ 
+    typeDefs, 
+    resolvers, 
+    context: authMiddleware 
+  });
+
+  // Initialize server
+  await server.start();
+
+  // integrate Apollo server with Express as middleware
+  server.applyMiddleware({ app });
+
+  // log where we can go to test our GQL API
+  console.log('Use GraphQL at http://localhost:' + PORT + server.graphqlPath);
+};
+
+// Serve static assets
+if (process.env.NODE_ENV === 'production') {
+  app.use(express.static(path.join(__dirname, '../client/build')));
+}
+
+// remove comment from below at time of deployment!
+// app.get('*', (req, res) => {
+//   res.sendFile(path.join(__dirname, '../client/build/index.html'));
+// });
+
+// Initialize server
+startServer();
+
+app.use(express.urlencoded({ extended: false }));
+app.use(express.json());
+
+db.once('open', () => {
+  app.listen(PORT, () => {
+    console.log('Server is served! Now listening on ' + PORT + '!');
+  });
+});
+  `
+}
+
+// generating .gitignore for root and server directories
+
+export const generateGitignore = () => {
+  return `# Logs
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+lerna-debug.log*
+
+# Diagnostic reports (https://nodejs.org/api/report.html)
+report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
+
+# Runtime data
+pids
+*.pid
+*.seed
+*.pid.lock
+
+# Directory for instrumented libs generated by jscoverage/JSCover
+lib-cov
+
+# Coverage directory used by tools like istanbul
+coverage
+*.lcov
+
+# nyc test coverage
+.nyc_output
+
+# Grunt intermediate storage (https://gruntjs.com/creating-plugins#storing-task-files)
+.grunt
+
+# Bower dependency directory (https://bower.io/)
+bower_components
+
+# node-waf configuration
+.lock-wscript
+
+# Compiled binary addons (https://nodejs.org/api/addons.html)
+build/Release
+
+# Dependency directories
+node_modules/
+jspm_packages/
+
+# TypeScript v1 declaration files
+typings/
+
+# TypeScript cache
+*.tsbuildinfo
+
+# Optional npm cache directory
+.npm
+
+# Optional eslint cache
+.eslintcache
+
+# Microbundle cache
+.rpt2_cache/
+.rts2_cache_cjs/
+.rts2_cache_es/
+.rts2_cache_umd/
+
+# Optional REPL history
+.node_repl_history
+
+# Output of 'npm pack'
+*.tgz
+
+# Yarn Integrity file
+.yarn-integrity
+
+# dotenv environment variables file
+.env
+.env.test
+
+# parcel-bundler cache (https://parceljs.org/)
+.cache
+
+# Next.js build output
+.next
+
+# Nuxt.js build / generate output
+.nuxt
+dist
+
+# Gatsby files
+.cache/
+# Comment in the public line in if your project uses Gatsby and *not* Next.js
+# https://nextjs.org/blog/next-9-1#public-directory-support
+# public
+
+# vuepress build output
+.vuepress/dist
+
+# Serverless directories
+.serverless/
+
+# FuseBox cache
+.fusebox/
+
+# DynamoDB Local files
+.dynamodb/
+
+# TernJS port file
+.tern-port
+  `;
+};
+
+export const generateUser = () => {
+  return `const { Schema, model } = require('mongoose');
+const bcrypt = require('bcrypt');
+
+const userSchema = new Schema(
+  {
+    username: {
+      type: String,
+      required: true,
+      unique: true,
+      trim: true
+    },
+    email: {
+      type: String,
+      required: true,
+      unique: true,
+      // verify email is valid using RFC 5322 official standard
+      match: [/(?:[a-z0-9!#$%&'*+/=?^_\`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_\`{|}~-]+)*|"(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21\\x23-\\x5b\\x5d-\\x7f]|\\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])*")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21-\\x5a\\x53-\\x7f]|\\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])+)\\])/g, 'Please enter a valid email address.']
+    },
+    password: {
+      type: String,
+      required: true,
+      minlength: 8
+    },
+  },
+  {
+    toJSON: {
+      virtuals: true
+    }
+  }
+);
+
+// create password with pre-save middleware
+userSchema.pre('save', async function(next) {
+  if (this.isNew || this.isModified('password')) {
+    const saltRounds = 15;
+    this.password = await bcrypt.hash(this.password, saltRounds);
+  }
+
+  next();
+});
+
+// compare the incoming password with the hashed password
+userSchema.methods.isCorrectPassword = async function(password) {
+  return bcrypt.compare(password, this.password);
+};
+
+const User = model('User', userSchema);
+
+module.exports = User;  
+  `
+}
+
+export const generateEnv = () => {
+  return `SKIP_PREFLIGHT_CHECK=true`
+}
+
+export function generateJSON(data) {
+  return data
+}
+
+export const generateModelsIndex = () => {
+  return `const User = require('./User');
+
+module.exports = { User }  
+  `
+}
+
+export const generateResolvers = () => {
+  return `const { User } = require("../models");
+const { AuthenticationError } = require("apollo-server-express");
+const { signToken } = require("../utils/auth");
+
+const resolvers = {
+  Query: {
+    me: async (parent, args, context) => {
+      if (context.user) {
+        const userData = await User.findOne({ _id: context.user._id })
+          .select('-__v -password')
+    
+        return userData;
+      }
+    
+      throw new AuthenticationError('Please log in first.');
+    },
+    // get all users
+    users: async () => {
+      return User.find()
+        .select("-__v -password")
+    },
+    // get user by username
+    user: async (parent, { username }) => {
+      return User.findOne({ username })
+        .select("-__v -password")
+    },
+  },
+  Mutation: {
+    addUser: async (parent, args) => {
+      const user = await User.create(args);
+      const token = signToken(user);
+
+      return { token, user };
+    },
+    login: async (parent, { email, password }) => {
+      const user = await User.findOne({ email });
+
+      if (!user) {
+        throw new AuthenticationError("Those credentials are incorrect.");
+      }
+
+      const correctPassword = await user.isCorrectPassword(password);
+
+      if (!correctPassword) {
+        throw new AuthenticationError("Incorrect credentials");
+      }
+
+      const token = signToken(user);
+      return { token, user };
+    },
+  },
+};
+
+module.exports = resolvers;
+
+  
+  `
+}
+
+export const generateSchemasIndex = () => {
+  return `const typeDefs = require('./typeDefs');
+const resolvers = require('./resolvers');
+
+module.exports = { typeDefs, resolvers };
+  `
+}
+
+export const generateSeeds = () => {
+  return `const faker = require('faker/locale/en_US');
+
+const db = require('../config/connection');
+const { User } = require('../models');
+
+db.once('open', async () => {
+  await User.deleteMany({});
+
+  // create data for faked users
+  const userData = [];
+
+  for (let i = 0; i < 20; i += 1) {
+    const username = faker.internet.userName();
+    const email = faker.internet.email(username);
+    const password = faker.internet.password();
+
+    userData.push({ username, email, password });
+  }
+
+  await User.collection.insertMany(userData);
+
+  console.log('Your seeds have been planted! 🌱');
+  process.exit(0);
+});  
+  `
+}
+
+export const generateTypeDefs = () => {
+  return `// import gql function (tagged template)
+const { gql } = require("apollo-server-express");
+
+// create typeDefs
+const typeDefs = gql\`
+
+  type User {
+    _id: ID
+    username: String
+    email: String
+  }
+
+  type Query {
+    me: User
+    users: [User]
+    user(username: String!): User
+  }
+
+  type Mutation {
+    login(email: String!, password: String!): Auth
+    addUser(username: String!, email: String!, password: String!): Auth
+  }
+
+  type Auth {
+    token: ID!
+    user: User
+  }
+
+\`;
+
+// export typeDefs
+module.exports = typeDefs;
+  
+  
+  `
+}
