+++ utils/setupQuestions.js
@@ -1,0 +1,15 @@
+export const setupQuestions = [
+  {
+    type: "input",
+    name: "title",
+    message: "What is the title of your project?",
+    validate: (titleInput) => {
+      if (titleInput) {
+        return true;
+      } else {
+        console.log("Please enter a project title.");
+        return false;
+      }
+    },
+  },
+];
